requirements.txt
requests
python-dotenv
pandas
openpyxl
supabase
pyjwt
pyairtable
